added failsafe

+++ datasetEvening/datasetEvening.rb
@@ -2,7 +2,7 @@ def smallestNumFiles()
     numFiles = []
     labels = Dir.glob("*")
     labels.each do |i|
-        # if i != "datasetEvening.rb"
+        if i != "datasetEvening.rb"
             Dir.chdir(i)
             fileCount = Dir.glob("*").count
             if fileCount != 0
@@ -12,29 +12,29 @@ def smallestNumFiles()
                 Dir.chdir("..")
                 Dir.rmdir(i)
             end
-        # end
+        end
     end
 
     return numFiles.min()
 end
 
-def datasetEvening(mapp)
-    Dir.chdir(mapp)
+def datasetEvening()
+    # Dir.chdir(mapp)
     targetFiles = smallestNumFiles()
     
     labels = Dir.glob("*")
     labels.each do |i|
-        # if i != "datasetEvening.rb"
+        if i != "datasetEvening.rb"
             Dir.chdir(i)
             while Dir.glob("*").count > targetFiles
                 i = Dir.glob("*")
                 len = i.length-1
                 File.delete(i[rand(0..len)])
             end
-            p i + ": Done"
+            # p i + ": Done"
             Dir.chdir("..")
-        # end
+        end
     end
 end
 
-datasetEvening("datasets")
+datasetEvening()
